test-data-generator/requirements.txt
PySocks
paho-mqtt
protobuf
xmltodict
pandas
schedule
requests